pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

# ---------------------------------------------------------------------------
# Project metadata
# ---------------------------------------------------------------------------
[project]
name = "llm-toolkit-schema"
version = "1.1.0"
description = "The foundational shared event schema for the LLM Developer Toolkit"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "LLM Toolkit Team" }]
keywords = ["llm", "schema", "opentelemetry", "events", "observability"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Typing :: Typed",
]
requires-python = ">=3.9"

# Core has ZERO required dependencies — stdlib only.
dependencies = []

[project.urls]
"Homepage" = "https://github.com/llm-toolkit/llm-toolkit-schema"
"Documentation" = "https://llm-toolkit-schema.readthedocs.io"
"Source" = "https://github.com/llm-toolkit/llm-toolkit-schema"
"Bug Tracker" = "https://github.com/llm-toolkit/llm-toolkit-schema/issues"
"Changelog" = "https://llm-toolkit-schema.readthedocs.io/en/latest/changelog.html"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.5",     # parallel test execution
    "pytest-benchmark>=4.0",  # performance regression tests
    "ruff>=0.4.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "hypothesis>=6.100",      # property-based testing
    "pydantic>=2.7",           # Phase 2: Pydantic model layer
    "httpx>=0.27",             # Phase 4: async webhook / OTLP HTTP
    "pytest-asyncio>=0.23",    # Phase 4: async test support
    "jsonschema>=4.21",        # Phase 5: JSON Schema validation (optional dep)
]
pydantic = ["pydantic>=2.7"]
otel    = ["opentelemetry-sdk>=1.24"]
http    = ["httpx>=0.27"]
jsonschema = ["jsonschema>=4.21"]
# Phase 7 enterprise integrations
datadog = []   # DatadogExporter uses stdlib HTTP, no extra deps required
kafka   = ["kafka-python>=2.0"]
# Phase 8 ecosystem integrations
langchain   = ["langchain-core>=0.2"]
llamaindex  = ["llama-index-core>=0.10"]
# Convenience bundle
all = [
    "pydantic>=2.7",
    "opentelemetry-sdk>=1.24",
    "httpx>=0.27",
    "jsonschema>=4.21",
    "kafka-python>=2.0",
    "langchain-core>=0.2",
    "llama-index-core>=0.10",
]
docs = [
    "sphinx>=7.0",
    "pydata-sphinx-theme>=0.15",
]

[project.scripts]
llm-toolkit-schema = "llm_toolkit_schema._cli:main"

# ---------------------------------------------------------------------------
# Hatch build
# ---------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["llm_toolkit_schema"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths     = ["tests"]
addopts       = [
    "--strict-markers",
    "--tb=short",
    "-ra",
    "--cov=llm_toolkit_schema",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=100",
]
markers = [
    "unit: fast, isolated unit tests",
    "integration: tests that cross module boundaries",
    "perf: performance benchmarks",
    "security: security-focused tests",
]
filterwarnings = ["error"]   # treat all warnings as errors

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
branch   = true
source   = ["llm_toolkit_schema"]
omit     = ["llm_toolkit_schema/_cli.py"]   # CLI scaffolded but not in Phase 1 scope

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "^\\s*\\.\\.\\.$",              # abstract method bodies
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true
fail_under   = 100

# ---------------------------------------------------------------------------
# Ruff — linting + formatting  (replaces flake8, isort, pyupgrade, bandit-lite)
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py39"
line-length    = 100
src            = ["llm_toolkit_schema", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",    # pycodestyle
    "F",         # pyflakes
    "I",         # isort
    "N",         # pep8-naming
    "UP",        # pyupgrade
    "B",         # flake8-bugbear
    "C4",        # flake8-comprehensions
    "SIM",       # flake8-simplify
    "TCH",       # flake8-type-checking
    "PTH",       # flake8-use-pathlib
    "RUF",       # Ruff-native rules
    "S",         # flake8-bandit (security)
    "ANN",       # flake8-annotations
    "D",         # pydocstyle
    "ERA",       # eradicate (commented-out code)
    "PL",        # pylint subset
    "PERF",      # perflint
    "TRY",       # tryceratops (exception anti-patterns)
]
ignore = [
    "D105",   # missing docstring in magic method (too verbose)
    "D107",   # missing docstring in __init__
    "ANN101", # missing type annotation for `self`
    "ANN102", # missing type annotation for `cls`
    "TRY003", # avoid long messages in raise — we intentionally add context
    "S101",   # use of assert — fine in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ANN", "D", "PLR2004"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["llm_toolkit_schema"]

# ---------------------------------------------------------------------------
# Mypy — strict type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version         = "3.9"
strict                 = true
warn_unused_ignores    = true
warn_redundant_casts   = true
warn_return_any        = true
disallow_any_generics  = true
disallow_untyped_defs  = true
no_implicit_reexport   = true
show_error_codes       = true
pretty                 = true
packages               = ["llm_toolkit_schema"]
